Refactor kms and add toolings unit tests
